refac block_propose_service: publish event
